+++ src/TrashLib.Tests/Sonarr/SonarrConfigurationTest.cs
@@ -1,47 +1,34 @@
-using Autofac;
 using FluentAssertions;
 using FluentValidation;
 using NUnit.Framework;
-using TrashLib.Config;
-using TrashLib.Services.Sonarr;
+using Recyclarr.TestLibrary;
 using TrashLib.Services.Sonarr.Config;
 
 namespace TrashLib.Tests.Sonarr;
 
 [TestFixture]
 [Parallelizable(ParallelScope.All)]
-public class SonarrConfigurationTest
+public class SonarrConfigurationTest : IntegrationFixture
 {
-    private IContainer _container = default!;
-
-    [OneTimeSetUp]
-    public void Setup()
-    {
-        var builder = new ContainerBuilder();
-        builder.RegisterModule<ConfigAutofacModule>();
-        builder.RegisterModule<SonarrAutofacModule>();
-        _container = builder.Build();
-    }
-
     [Test]
     public void Validation_fails_for_all_missing_required_properties()
     {
         // default construct which should yield default values (invalid) for all required properties
         var config = new SonarrConfiguration
         {
+            BaseUrl = "valid",
+            ApiKey = "valid",
             // validation is only applied to actual release profile elements. Not if it's empty.
             ReleaseProfiles = new[] {new ReleaseProfileConfig()}
         };
 
-        var validator = _container.Resolve<IValidator<SonarrConfiguration>>();
+        var validator = ServiceLocator.Resolve<IValidator<SonarrConfiguration>>();
 
         var result = validator.Validate(config);
 
         var messages = new SonarrValidationMessages();
         var expectedErrorMessageSubstrings = new[]
         {
-            messages.ApiKey,
-            messages.BaseUrl,
             messages.ReleaseProfileTrashIds
         };
 
@@ -63,7 +50,7 @@ public class SonarrConfigurationTest
             }
         };
 
-        var validator = _container.Resolve<IValidator<SonarrConfiguration>>();
+        var validator = ServiceLocator.Resolve<IValidator<SonarrConfiguration>>();
         var result = validator.Validate(config);
 
         result.IsValid.Should().BeTrue();