+++ src/TrashLib/Config/Services/IServiceValidationMessages.cs
@@ -1,0 +1,8 @@
+namespace TrashLib.Config.Services;
+
+public interface IServiceValidationMessages
+{
+    string BaseUrl { get; }
+    string ApiKey { get; }
+    string CustomFormatTrashIds { get; }
+}