+++ src/TrashLib/Config/Services/ServiceConfiguration.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 
 namespace TrashLib.Config.Services;
 
-public abstract class ServiceConfiguration : IServiceConfiguration
+public class ServiceConfiguration : IServiceConfiguration
 {
     public string BaseUrl { get; init; } = "";
     public string ApiKey { get; init; } = "";