+++ src/TrashLib/Services/Radarr/Config/IRadarrValidationMessages.cs
@@ -2,9 +2,6 @@ namespace TrashLib.Services.Radarr.Config;
 
 public interface IRadarrValidationMessages
 {
-    string BaseUrl { get; }
-    string ApiKey { get; }
-    string CustomFormatTrashIds { get; }
     string QualityProfileName { get; }
     string QualityDefinitionType { get; }
 }