+++ src/TrashLib/Services/Sonarr/Config/ISonarrValidationMessages.cs
@@ -2,7 +2,5 @@ namespace TrashLib.Services.Sonarr.Config;
 
 public interface ISonarrValidationMessages
 {
-    string BaseUrl { get; }
-    string ApiKey { get; }
     string ReleaseProfileTrashIds { get; }
 }